# java/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /games:
    get:
      tags:
      - trivia-controller
      operationId: listGames
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameResponseDto'
    post:
      tags:
      - trivia-controller
      operationId: createGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequestDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponseDto'
  /games/{gameId}/players/{playerId}/start:
    post:
      tags:
      - trivia-controller
      operationId: startGame
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponseDto'
  /games/{gameId}/players/{playerId}/playTurn:
    post:
      tags:
      - trivia-controller
      operationId: playTurn
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponseDto'
  /games/{gameId}/players/{playerId}/join:
    post:
      tags:
      - trivia-controller
      operationId: addPlayerToGame
      parameters:
      - name: gameId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: playerId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponseDto'
components:
  schemas:
    CreateGameRequestDto:
      type: object
      properties:
        gameName:
          type: string
        creator:
          $ref: '#/components/schemas/UserDto'
      required:
        - gameName
        - creator
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
      required:
        - id
        - name
    GameResponseDto:
      type: object
      properties:
        id:
          type: integer
          nullable: false
          format: int32
        name:
          type: string
          nullable: false
        state:
          type: string
        creator:
          $ref: '#/components/schemas/UserDto'
        players:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        currentPlayer:
          $ref: '#/components/schemas/UserDto'
      required:
        - id
        - name
        - state
        - creator
        - players
        - currentPlayer